+++ lib/core/api/endpoints/post_endpoint.dart
@@ -8,4 +8,7 @@ class PostEndpoint {
 
   static String indicate(String postId) 
     => "${_env.apiHost}/v1/posts/$postId/indications";
+    
+  static String like(String postId) 
+    => "${_env.apiHost}/v1/posts/$postId/likes";
 }