+++ lib/core/api/repositories/post/post_repository.dart
@@ -1,6 +1,8 @@
 import 'package:dartz/dartz.dart';
+import 'package:http/http.dart';
 import 'package:music_playce_sdk/core/api/models/cursor.dart';
 import 'package:music_playce_sdk/core/api/models/posts/indication.model.dart';
+import 'package:music_playce_sdk/core/api/models/posts/like.model.dart';
 import 'package:music_playce_sdk/core/api/models/posts/posts.model.dart';
 
 /// A repository for deal with posts from the api
@@ -10,4 +12,8 @@ abstract class PostRepository {
 
   // TODO: put playback moment
   Future<Either<Exception, Indication>> indicate(String postId, String indicatedProfileId);
+
+  Future<Either<Exception, Like>> like(String postId);
+
+  Future<Either<Exception, Response>> dislike(String postId);
 }